requirements.txt
picamera
RPi.GPIO
adafruit-pca9685
rpi_ws281x